--- src/modules/UsersProfile/components/index.ts
@@ -1,1 +1,0 @@
-export { default as LoginForm } from './LoginForm'
