+++ src/store/index.ts
@@ -1,4 +1,3 @@
 export { useSavedStatsStore } from './saveStatStore'
-export { useUserAuthStore } from './userAuthStore'
 export { useUsersStore } from './usersStore'
 export { useGraphStore } from './graphStore'
